Scrub GH Image links (#305)
* Scrub GH Image links

When an image is pasted into an issue, or PR description or comment, the anchor link contains a JWT for the image. The host is `private-user-images.githubusercontent.com`. This link gets flagged by our credential scanner as possibly containing a user secret. Therefore, the import fails.

Add an extension method to scrub HTML from GH so that it can be imported into Azure DevOps. At the moment, this method only removes these links. If other content becomes problematic, we can extend.

* Apply suggestions from code review

Co-authored-by: David Pine <david.pine@microsoft.com>

---------

Co-authored-by: David Pine <david.pine@microsoft.com>